+++ feature/trip-planner/ui/src/commonMain/kotlin/xyz/ksharma/krail/trip/planner/ui/settings/story/OurStoryDestination.kt
@@ -1,4 +1,4 @@
-package xyz.ksharma.krail.trip.planner.ui.settings.about
+package xyz.ksharma.krail.trip.planner.ui.settings.story
 
 import androidx.compose.runtime.getValue
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
@@ -6,16 +6,16 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
 import org.koin.compose.viewmodel.koinViewModel
-import xyz.ksharma.krail.trip.planner.ui.navigation.AboutUsRoute
+import xyz.ksharma.krail.trip.planner.ui.navigation.OurStoryRoute
 
 internal fun NavGraphBuilder.aboutUsDestination(navController: NavHostController) {
-    composable<AboutUsRoute> {
+    composable<OurStoryRoute> {
 
-        val viewModel: AboutUsViewModel = koinViewModel<AboutUsViewModel>()
-        val aboutUsState by viewModel.uiState.collectAsStateWithLifecycle()
+        val viewModel: OurStoryViewModel = koinViewModel<OurStoryViewModel>()
+        val ourStoryState by viewModel.uiState.collectAsStateWithLifecycle()
 
-        AboutUsScreen(
-            state = aboutUsState,
+        OurStoryScreen(
+            state = ourStoryState,
             onBackClick = { navController.popBackStack() }
         )
     }
